requirements.txt
autopep8==1.5.7
certifi==2021.10.8
charset-normalizer==2.0.7
idna==3.3
pycodestyle==2.8.0
PyYAML==6.0
requests==2.26.0
toml==0.10.2
urllib3==1.26.7
